+++ datafusion/core/src/physical_optimizer/projection_pushdown.rs
@@ -24,31 +24,18 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use crate::error::Result;
-use crate::physical_plan::coalesce_partitions::CoalescePartitionsExec;
-use crate::physical_plan::filter::FilterExec;
-use crate::physical_plan::joins::utils::{ColumnIndex, JoinFilter};
-use crate::physical_plan::joins::{
-    CrossJoinExec, HashJoinExec, NestedLoopJoinExec, SortMergeJoinExec,
-    SymmetricHashJoinExec,
-};
-
 use crate::physical_plan::projection::ProjectionExec;
+use crate::physical_plan::ExecutionPlan;
 
-use crate::physical_plan::{ExecutionPlan, ExecutionPlanProperties};
-
-use arrow_schema::SchemaRef;
 use datafusion_common::config::ConfigOptions;
 use datafusion_common::tree_node::{
     Transformed, TransformedResult, TreeNode, TreeNodeRecursion,
 };
-use datafusion_common::{internal_err, JoinSide};
 use datafusion_physical_expr::expressions::{Column, Literal};
-use datafusion_physical_expr::{utils::collect_columns, PhysicalExpr, PhysicalExprRef};
-use datafusion_physical_plan::joins::utils::{JoinOn, JoinOnRef};
+use datafusion_physical_expr::PhysicalExpr;
 
-use datafusion_physical_expr_common::sort_expr::LexOrdering;
 use datafusion_physical_optimizer::PhysicalOptimizerRule;
-use itertools::Itertools;
+use datafusion_physical_plan::projection_utils::update_expr;
 
 /// This rule inspects [`ProjectionExec`]'s in the given physical plan and tries to
 /// remove or swap with its child.
@@ -107,25 +94,6 @@ pub fn remove_unnecessary_projections(
             } else {
                 Ok(Transformed::no(plan))
             };
-        } else if input.is::<CoalescePartitionsExec>() {
-            try_swapping_with_coalesce_partitions(projection)?
-        } else if let Some(filter) = input.downcast_ref::<FilterExec>() {
-            filter
-                .try_swapping_with_projection(projection)?
-                .map_or_else(
-                    || try_embed_projection(projection, filter),
-                    |e| Ok(Some(e)),
-                )?
-        } else if let Some(hash_join) = input.downcast_ref::<HashJoinExec>() {
-            try_pushdown_through_hash_join(projection, hash_join)?
-        } else if let Some(cross_join) = input.downcast_ref::<CrossJoinExec>() {
-            try_swapping_with_cross_join(projection, cross_join)?
-        } else if let Some(nl_join) = input.downcast_ref::<NestedLoopJoinExec>() {
-            try_pushdown_through_nested_loop_join(projection, nl_join)?
-        } else if let Some(sm_join) = input.downcast_ref::<SortMergeJoinExec>() {
-            try_swapping_with_sort_merge_join(projection, sm_join)?
-        } else if let Some(sym_join) = input.downcast_ref::<SymmetricHashJoinExec>() {
-            try_swapping_with_sym_hash_join(projection, sym_join)?
         } else {
             projection
                 .input()
@@ -190,434 +158,6 @@ fn is_expr_trivial(expr: &Arc<dyn PhysicalExpr>) -> bool {
         || expr.as_any().downcast_ref::<Literal>().is_some()
 }
 
-/// Tries to swap `projection` with its input, which is known to be a
-/// [`CoalescePartitionsExec`]. If possible, performs the swap and returns
-/// [`CoalescePartitionsExec`] as the top plan. Otherwise, returns `None`.
-fn try_swapping_with_coalesce_partitions(
-    projection: &ProjectionExec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // If the projection does not narrow the schema, we should not try to push it down:
-    if projection.expr().len() >= projection.input().schema().fields().len() {
-        return Ok(None);
-    }
-    // CoalescePartitionsExec always has a single child, so zero indexing is safe.
-    make_with_child(projection, projection.input().children()[0])
-        .map(|e| Some(Arc::new(CoalescePartitionsExec::new(e)) as _))
-}
-
-trait EmbeddedProjection: ExecutionPlan + Sized {
-    fn with_projection(&self, projection: Option<Vec<usize>>) -> Result<Self>;
-}
-
-impl EmbeddedProjection for HashJoinExec {
-    fn with_projection(&self, projection: Option<Vec<usize>>) -> Result<Self> {
-        self.with_projection(projection)
-    }
-}
-
-impl EmbeddedProjection for NestedLoopJoinExec {
-    fn with_projection(&self, projection: Option<Vec<usize>>) -> Result<Self> {
-        self.with_projection(projection)
-    }
-}
-
-impl EmbeddedProjection for FilterExec {
-    fn with_projection(&self, projection: Option<Vec<usize>>) -> Result<Self> {
-        self.with_projection(projection)
-    }
-}
-
-/// Some projection can't be pushed down left input or right input of hash join because filter or on need may need some columns that won't be used in later.
-/// By embed those projection to hash join, we can reduce the cost of build_batch_from_indices in hash join (build_batch_from_indices need to can compute::take() for each column) and avoid unnecessary output creation.
-fn try_embed_projection<Exec: EmbeddedProjection + 'static>(
-    projection: &ProjectionExec,
-    execution_plan: &Exec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // Collect all column indices from the given projection expressions.
-    let projection_index = collect_column_indices(projection.expr());
-
-    if projection_index.is_empty() {
-        return Ok(None);
-    };
-
-    // If the projection indices is the same as the input columns, we don't need to embed the projection to hash join.
-    // Check the projection_index is 0..n-1 and the length of projection_index is the same as the length of execution_plan schema fields.
-    if projection_index.len() == projection_index.last().unwrap() + 1
-        && projection_index.len() == execution_plan.schema().fields().len()
-    {
-        return Ok(None);
-    }
-
-    let new_execution_plan =
-        Arc::new(execution_plan.with_projection(Some(projection_index.to_vec()))?);
-
-    // Build projection expressions for update_expr. Zip the projection_index with the new_execution_plan output schema fields.
-    let embed_project_exprs = projection_index
-        .iter()
-        .zip(new_execution_plan.schema().fields())
-        .map(|(index, field)| {
-            (
-                Arc::new(Column::new(field.name(), *index)) as Arc<dyn PhysicalExpr>,
-                field.name().to_owned(),
-            )
-        })
-        .collect::<Vec<_>>();
-
-    let mut new_projection_exprs = Vec::with_capacity(projection.expr().len());
-
-    for (expr, alias) in projection.expr() {
-        // update column index for projection expression since the input schema has been changed.
-        let Some(expr) = update_expr(expr, embed_project_exprs.as_slice(), false)? else {
-            return Ok(None);
-        };
-        new_projection_exprs.push((expr, alias.clone()));
-    }
-    // Old projection may contain some alias or expression such as `a + 1` and `CAST('true' AS BOOLEAN)`, but our projection_exprs in hash join just contain column, so we need to create the new projection to keep the original projection.
-    let new_projection = Arc::new(ProjectionExec::try_new(
-        new_projection_exprs,
-        Arc::clone(&new_execution_plan) as _,
-    )?);
-    if is_projection_removable(&new_projection) {
-        Ok(Some(new_execution_plan))
-    } else {
-        Ok(Some(new_projection))
-    }
-}
-
-/// Collect all column indices from the given projection expressions.
-fn collect_column_indices(exprs: &[(Arc<dyn PhysicalExpr>, String)]) -> Vec<usize> {
-    // Collect indices and remove duplicates.
-    let mut indices = exprs
-        .iter()
-        .flat_map(|(expr, _)| collect_columns(expr))
-        .map(|x| x.index())
-        .collect::<std::collections::HashSet<_>>()
-        .into_iter()
-        .collect::<Vec<_>>();
-    indices.sort();
-    indices
-}
-
-struct JoinData {
-    projected_left_child: ProjectionExec,
-    projected_right_child: ProjectionExec,
-    join_filter: Option<JoinFilter>,
-    join_on: JoinOn,
-}
-
-fn try_pushdown_through_join(
-    projection: &ProjectionExec,
-    join_left: &Arc<dyn ExecutionPlan>,
-    join_right: &Arc<dyn ExecutionPlan>,
-    join_on: JoinOnRef,
-    schema: SchemaRef,
-    filter: Option<&JoinFilter>,
-) -> Result<Option<JoinData>> {
-    // Convert projected expressions to columns. We can not proceed if this is not possible.
-    let Some(projection_as_columns) = physical_to_column_exprs(projection.expr()) else {
-        return Ok(None);
-    };
-
-    let (far_right_left_col_ind, far_left_right_col_ind) =
-        join_table_borders(join_left.schema().fields().len(), &projection_as_columns);
-
-    if !join_allows_pushdown(
-        &projection_as_columns,
-        &schema,
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-    ) {
-        return Ok(None);
-    }
-
-    let new_filter = if let Some(filter) = filter {
-        match update_join_filter(
-            &projection_as_columns[0..=far_right_left_col_ind as _],
-            &projection_as_columns[far_left_right_col_ind as _..],
-            filter,
-            join_left.schema().fields().len(),
-        ) {
-            Some(updated_filter) => Some(updated_filter),
-            None => return Ok(None),
-        }
-    } else {
-        None
-    };
-
-    let Some(new_on) = update_join_on(
-        &projection_as_columns[0..=far_right_left_col_ind as _],
-        &projection_as_columns[far_left_right_col_ind as _..],
-        join_on,
-        join_left.schema().fields().len(),
-    ) else {
-        return Ok(None);
-    };
-
-    let (new_left, new_right) = new_join_children(
-        &projection_as_columns,
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-        join_left,
-        join_right,
-    )?;
-
-    Ok(Some(JoinData {
-        projected_left_child: new_left,
-        projected_right_child: new_right,
-        join_filter: new_filter,
-        join_on: new_on,
-    }))
-}
-
-/// Tries to push `projection` down through `nested_loop_join`. If possible, performs the
-/// pushdown and returns a new [`NestedLoopJoinExec`] as the top plan which has projections
-/// as its children. Otherwise, returns `None`.
-fn try_pushdown_through_nested_loop_join(
-    projection: &ProjectionExec,
-    nl_join: &NestedLoopJoinExec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // TODO: currently if there is projection in NestedLoopJoinExec, we can't push down projection to left or right input. Maybe we can pushdown the mixed projection later.
-    if nl_join.contains_projection() {
-        return Ok(None);
-    }
-
-    if let Some(JoinData {
-        projected_left_child,
-        projected_right_child,
-        join_filter,
-        ..
-    }) = try_pushdown_through_join(
-        projection,
-        nl_join.left(),
-        nl_join.right(),
-        &[],
-        nl_join.schema(),
-        nl_join.filter(),
-    )? {
-        Ok(Some(Arc::new(NestedLoopJoinExec::try_new(
-            Arc::new(projected_left_child),
-            Arc::new(projected_right_child),
-            join_filter,
-            nl_join.join_type(),
-            // Returned early if projection is not None
-            None,
-        )?)))
-    } else {
-        try_embed_projection(projection, nl_join)
-    }
-}
-
-/// Tries to push `projection` down through `hash_join`. If possible, performs the
-/// pushdown and returns a new [`HashJoinExec`] as the top plan which has projections
-/// as its children. Otherwise, returns `None`.
-fn try_pushdown_through_hash_join(
-    projection: &ProjectionExec,
-    hash_join: &HashJoinExec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // TODO: currently if there is projection in HashJoinExec, we can't push down projection to left or right input. Maybe we can pushdown the mixed projection later.
-    if hash_join.contains_projection() {
-        return Ok(None);
-    }
-
-    if let Some(JoinData {
-        projected_left_child,
-        projected_right_child,
-        join_filter,
-        join_on,
-    }) = try_pushdown_through_join(
-        projection,
-        hash_join.left(),
-        hash_join.right(),
-        hash_join.on(),
-        hash_join.schema(),
-        hash_join.filter(),
-    )? {
-        Ok(Some(Arc::new(HashJoinExec::try_new(
-            Arc::new(projected_left_child),
-            Arc::new(projected_right_child),
-            join_on,
-            join_filter,
-            hash_join.join_type(),
-            // Returned early if projection is not None
-            None,
-            *hash_join.partition_mode(),
-            hash_join.null_equals_null,
-        )?)))
-    } else {
-        try_embed_projection(projection, hash_join)
-    }
-}
-
-/// Tries to swap the projection with its input [`CrossJoinExec`]. If it can be done,
-/// it returns the new swapped version having the [`CrossJoinExec`] as the top plan.
-/// Otherwise, it returns None.
-fn try_swapping_with_cross_join(
-    projection: &ProjectionExec,
-    cross_join: &CrossJoinExec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // Convert projected PhysicalExpr's to columns. If not possible, we cannot proceed.
-    let Some(projection_as_columns) = physical_to_column_exprs(projection.expr()) else {
-        return Ok(None);
-    };
-
-    let (far_right_left_col_ind, far_left_right_col_ind) = join_table_borders(
-        cross_join.left().schema().fields().len(),
-        &projection_as_columns,
-    );
-
-    if !join_allows_pushdown(
-        &projection_as_columns,
-        &cross_join.schema(),
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-    ) {
-        return Ok(None);
-    }
-
-    let (new_left, new_right) = new_join_children(
-        &projection_as_columns,
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-        cross_join.left(),
-        cross_join.right(),
-    )?;
-
-    Ok(Some(Arc::new(CrossJoinExec::new(
-        Arc::new(new_left),
-        Arc::new(new_right),
-    ))))
-}
-
-/// Tries to swap the projection with its input [`SortMergeJoinExec`]. If it can be done,
-/// it returns the new swapped version having the [`SortMergeJoinExec`] as the top plan.
-/// Otherwise, it returns None.
-fn try_swapping_with_sort_merge_join(
-    projection: &ProjectionExec,
-    sm_join: &SortMergeJoinExec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // Convert projected PhysicalExpr's to columns. If not possible, we cannot proceed.
-    let Some(projection_as_columns) = physical_to_column_exprs(projection.expr()) else {
-        return Ok(None);
-    };
-
-    let (far_right_left_col_ind, far_left_right_col_ind) = join_table_borders(
-        sm_join.left().schema().fields().len(),
-        &projection_as_columns,
-    );
-
-    if !join_allows_pushdown(
-        &projection_as_columns,
-        &sm_join.schema(),
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-    ) {
-        return Ok(None);
-    }
-
-    let Some(new_on) = update_join_on(
-        &projection_as_columns[0..=far_right_left_col_ind as _],
-        &projection_as_columns[far_left_right_col_ind as _..],
-        sm_join.on(),
-        sm_join.left().schema().fields().len(),
-    ) else {
-        return Ok(None);
-    };
-
-    let (new_left, new_right) = new_join_children(
-        &projection_as_columns,
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-        sm_join.children()[0],
-        sm_join.children()[1],
-    )?;
-
-    Ok(Some(Arc::new(SortMergeJoinExec::try_new(
-        Arc::new(new_left),
-        Arc::new(new_right),
-        new_on,
-        sm_join.filter.clone(),
-        sm_join.join_type,
-        sm_join.sort_options.clone(),
-        sm_join.null_equals_null,
-    )?)))
-}
-
-/// Tries to swap the projection with its input [`SymmetricHashJoinExec`]. If it can be done,
-/// it returns the new swapped version having the [`SymmetricHashJoinExec`] as the top plan.
-/// Otherwise, it returns None.
-fn try_swapping_with_sym_hash_join(
-    projection: &ProjectionExec,
-    sym_join: &SymmetricHashJoinExec,
-) -> Result<Option<Arc<dyn ExecutionPlan>>> {
-    // Convert projected PhysicalExpr's to columns. If not possible, we cannot proceed.
-    let Some(projection_as_columns) = physical_to_column_exprs(projection.expr()) else {
-        return Ok(None);
-    };
-
-    let (far_right_left_col_ind, far_left_right_col_ind) = join_table_borders(
-        sym_join.left().schema().fields().len(),
-        &projection_as_columns,
-    );
-
-    if !join_allows_pushdown(
-        &projection_as_columns,
-        &sym_join.schema(),
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-    ) {
-        return Ok(None);
-    }
-
-    let Some(new_on) = update_join_on(
-        &projection_as_columns[0..=far_right_left_col_ind as _],
-        &projection_as_columns[far_left_right_col_ind as _..],
-        sym_join.on(),
-        sym_join.left().schema().fields().len(),
-    ) else {
-        return Ok(None);
-    };
-
-    let new_filter = if let Some(filter) = sym_join.filter() {
-        match update_join_filter(
-            &projection_as_columns[0..=far_right_left_col_ind as _],
-            &projection_as_columns[far_left_right_col_ind as _..],
-            filter,
-            sym_join.left().schema().fields().len(),
-        ) {
-            Some(updated_filter) => Some(updated_filter),
-            None => return Ok(None),
-        }
-    } else {
-        None
-    };
-
-    let (new_left, new_right) = new_join_children(
-        &projection_as_columns,
-        far_right_left_col_ind,
-        far_left_right_col_ind,
-        sym_join.left(),
-        sym_join.right(),
-    )?;
-
-    Ok(Some(Arc::new(SymmetricHashJoinExec::try_new(
-        Arc::new(new_left),
-        Arc::new(new_right),
-        new_on,
-        new_filter,
-        sym_join.join_type(),
-        sym_join.null_equals_null(),
-        sym_join
-            .right()
-            .output_ordering()
-            .map(|p| LexOrdering::new(p.to_vec())),
-        sym_join
-            .left()
-            .output_ordering()
-            .map(|p| LexOrdering::new(p.to_vec())),
-        sym_join.partition_mode(),
-    )?)))
-}
-
 /// Compare the inputs and outputs of the projection. All expressions must be
 /// columns without alias, and projection does not change the order of fields.
 /// For example, if the input schema is `a, b`, `SELECT a, b` is removable,
@@ -630,357 +170,6 @@ fn is_projection_removable(projection: &ProjectionExec) -> bool {
         };
         col.name() == alias && col.index() == idx
     }) && exprs.len() == projection.input().schema().fields().len()
-}
-
-/// The function operates in two modes:
-///
-/// 1) When `sync_with_child` is `true`:
-///
-///    The function updates the indices of `expr` if the expression resides
-///    in the input plan. For instance, given the expressions `a@1 + b@2`
-///    and `c@0` with the input schema `c@2, a@0, b@1`, the expressions are
-///    updated to `a@0 + b@1` and `c@2`.
-///
-/// 2) When `sync_with_child` is `false`:
-///
-///    The function determines how the expression would be updated if a projection
-///    was placed before the plan associated with the expression. If the expression
-///    cannot be rewritten after the projection, it returns `None`. For example,
-///    given the expressions `c@0`, `a@1` and `b@2`, and the [`ProjectionExec`] with
-///    an output schema of `a, c_new`, then `c@0` becomes `c_new@1`, `a@1` becomes
-///    `a@0`, but `b@2` results in `None` since the projection does not include `b`.
-fn update_expr(
-    expr: &Arc<dyn PhysicalExpr>,
-    projected_exprs: &[(Arc<dyn PhysicalExpr>, String)],
-    sync_with_child: bool,
-) -> Result<Option<Arc<dyn PhysicalExpr>>> {
-    #[derive(Debug, PartialEq)]
-    enum RewriteState {
-        /// The expression is unchanged.
-        Unchanged,
-        /// Some part of the expression has been rewritten
-        RewrittenValid,
-        /// Some part of the expression has been rewritten, but some column
-        /// references could not be.
-        RewrittenInvalid,
-    }
-
-    let mut state = RewriteState::Unchanged;
-
-    let new_expr = Arc::clone(expr)
-        .transform_up(|expr: Arc<dyn PhysicalExpr>| {
-            if state == RewriteState::RewrittenInvalid {
-                return Ok(Transformed::no(expr));
-            }
-
-            let Some(column) = expr.as_any().downcast_ref::<Column>() else {
-                return Ok(Transformed::no(expr));
-            };
-            if sync_with_child {
-                state = RewriteState::RewrittenValid;
-                // Update the index of `column`:
-                Ok(Transformed::yes(Arc::clone(
-                    &projected_exprs[column.index()].0,
-                )))
-            } else {
-                // default to invalid, in case we can't find the relevant column
-                state = RewriteState::RewrittenInvalid;
-                // Determine how to update `column` to accommodate `projected_exprs`
-                projected_exprs
-                    .iter()
-                    .enumerate()
-                    .find_map(|(index, (projected_expr, alias))| {
-                        projected_expr.as_any().downcast_ref::<Column>().and_then(
-                            |projected_column| {
-                                (column.name().eq(projected_column.name())
-                                    && column.index() == projected_column.index())
-                                .then(|| {
-                                    state = RewriteState::RewrittenValid;
-                                    Arc::new(Column::new(alias, index)) as _
-                                })
-                            },
-                        )
-                    })
-                    .map_or_else(
-                        || Ok(Transformed::no(expr)),
-                        |c| Ok(Transformed::yes(c)),
-                    )
-            }
-        })
-        .data();
-
-    new_expr.map(|e| (state == RewriteState::RewrittenValid).then_some(e))
-}
-
-/// Creates a new [`ProjectionExec`] instance with the given child plan and
-/// projected expressions.
-fn make_with_child(
-    projection: &ProjectionExec,
-    child: &Arc<dyn ExecutionPlan>,
-) -> Result<Arc<dyn ExecutionPlan>> {
-    ProjectionExec::try_new(projection.expr().to_vec(), Arc::clone(child))
-        .map(|e| Arc::new(e) as _)
-}
-
-/// Downcasts all the expressions in `exprs` to `Column`s. If any of the given
-/// expressions is not a `Column`, returns `None`.
-fn physical_to_column_exprs(
-    exprs: &[(Arc<dyn PhysicalExpr>, String)],
-) -> Option<Vec<(Column, String)>> {
-    exprs
-        .iter()
-        .map(|(expr, alias)| {
-            expr.as_any()
-                .downcast_ref::<Column>()
-                .map(|col| (col.clone(), alias.clone()))
-        })
-        .collect()
-}
-
-/// Returns the last index before encountering a column coming from the right table when traveling
-/// through the projection from left to right, and the last index before encountering a column
-/// coming from the left table when traveling through the projection from right to left.
-/// If there is no column in the projection coming from the left side, it returns (-1, ...),
-/// if there is no column in the projection coming from the right side, it returns (..., projection length).
-fn join_table_borders(
-    left_table_column_count: usize,
-    projection_as_columns: &[(Column, String)],
-) -> (i32, i32) {
-    let far_right_left_col_ind = projection_as_columns
-        .iter()
-        .enumerate()
-        .take_while(|(_, (projection_column, _))| {
-            projection_column.index() < left_table_column_count
-        })
-        .last()
-        .map(|(index, _)| index as i32)
-        .unwrap_or(-1);
-
-    let far_left_right_col_ind = projection_as_columns
-        .iter()
-        .enumerate()
-        .rev()
-        .take_while(|(_, (projection_column, _))| {
-            projection_column.index() >= left_table_column_count
-        })
-        .last()
-        .map(|(index, _)| index as i32)
-        .unwrap_or(projection_as_columns.len() as i32);
-
-    (far_right_left_col_ind, far_left_right_col_ind)
-}
-
-/// Tries to update the equi-join `Column`'s of a join as if the input of
-/// the join was replaced by a projection.
-fn update_join_on(
-    proj_left_exprs: &[(Column, String)],
-    proj_right_exprs: &[(Column, String)],
-    hash_join_on: &[(PhysicalExprRef, PhysicalExprRef)],
-    left_field_size: usize,
-) -> Option<Vec<(PhysicalExprRef, PhysicalExprRef)>> {
-    // TODO: Clippy wants the "map" call removed, but doing so generates
-    //       a compilation error. Remove the clippy directive once this
-    //       issue is fixed.
-    #[allow(clippy::map_identity)]
-    let (left_idx, right_idx): (Vec<_>, Vec<_>) = hash_join_on
-        .iter()
-        .map(|(left, right)| (left, right))
-        .unzip();
-
-    let new_left_columns = new_columns_for_join_on(&left_idx, proj_left_exprs, 0);
-    let new_right_columns =
-        new_columns_for_join_on(&right_idx, proj_right_exprs, left_field_size);
-
-    match (new_left_columns, new_right_columns) {
-        (Some(left), Some(right)) => Some(left.into_iter().zip(right).collect()),
-        _ => None,
-    }
-}
-
-/// This function generates a new set of columns to be used in a hash join
-/// operation based on a set of equi-join conditions (`hash_join_on`) and a
-/// list of projection expressions (`projection_exprs`).
-///
-/// Notes: Column indices in the projection expressions are based on the join schema,
-/// whereas the join on expressions are based on the join child schema. `column_index_offset`
-/// represents the offset between them.
-fn new_columns_for_join_on(
-    hash_join_on: &[&PhysicalExprRef],
-    projection_exprs: &[(Column, String)],
-    column_index_offset: usize,
-) -> Option<Vec<PhysicalExprRef>> {
-    let new_columns = hash_join_on
-        .iter()
-        .filter_map(|on| {
-            // Rewrite all columns in `on`
-            Arc::clone(*on)
-                .transform(|expr| {
-                    if let Some(column) = expr.as_any().downcast_ref::<Column>() {
-                        // Find the column in the projection expressions
-                        let new_column = projection_exprs
-                            .iter()
-                            .enumerate()
-                            .find(|(_, (proj_column, _))| {
-                                column.name() == proj_column.name()
-                                    && column.index() + column_index_offset
-                                        == proj_column.index()
-                            })
-                            .map(|(index, (_, alias))| Column::new(alias, index));
-                        if let Some(new_column) = new_column {
-                            Ok(Transformed::yes(Arc::new(new_column)))
-                        } else {
-                            // If the column is not found in the projection expressions,
-                            // it means that the column is not projected. In this case,
-                            // we cannot push the projection down.
-                            internal_err!(
-                                "Column {:?} not found in projection expressions",
-                                column
-                            )
-                        }
-                    } else {
-                        Ok(Transformed::no(expr))
-                    }
-                })
-                .data()
-                .ok()
-        })
-        .collect::<Vec<_>>();
-    (new_columns.len() == hash_join_on.len()).then_some(new_columns)
-}
-
-/// Tries to update the column indices of a [`JoinFilter`] as if the input of
-/// the join was replaced by a projection.
-fn update_join_filter(
-    projection_left_exprs: &[(Column, String)],
-    projection_right_exprs: &[(Column, String)],
-    join_filter: &JoinFilter,
-    left_field_size: usize,
-) -> Option<JoinFilter> {
-    let mut new_left_indices = new_indices_for_join_filter(
-        join_filter,
-        JoinSide::Left,
-        projection_left_exprs,
-        0,
-    )
-    .into_iter();
-    let mut new_right_indices = new_indices_for_join_filter(
-        join_filter,
-        JoinSide::Right,
-        projection_right_exprs,
-        left_field_size,
-    )
-    .into_iter();
-
-    // Check if all columns match:
-    (new_right_indices.len() + new_left_indices.len()
-        == join_filter.column_indices().len())
-    .then(|| {
-        JoinFilter::new(
-            Arc::clone(join_filter.expression()),
-            join_filter
-                .column_indices()
-                .iter()
-                .map(|col_idx| ColumnIndex {
-                    index: if col_idx.side == JoinSide::Left {
-                        new_left_indices.next().unwrap()
-                    } else {
-                        new_right_indices.next().unwrap()
-                    },
-                    side: col_idx.side,
-                })
-                .collect(),
-            Arc::clone(join_filter.schema()),
-        )
-    })
-}
-
-/// This function determines and returns a vector of indices representing the
-/// positions of columns in `projection_exprs` that are involved in `join_filter`,
-/// and correspond to a particular side (`join_side`) of the join operation.
-///
-/// Notes: Column indices in the projection expressions are based on the join schema,
-/// whereas the join filter is based on the join child schema. `column_index_offset`
-/// represents the offset between them.
-fn new_indices_for_join_filter(
-    join_filter: &JoinFilter,
-    join_side: JoinSide,
-    projection_exprs: &[(Column, String)],
-    column_index_offset: usize,
-) -> Vec<usize> {
-    join_filter
-        .column_indices()
-        .iter()
-        .filter(|col_idx| col_idx.side == join_side)
-        .filter_map(|col_idx| {
-            projection_exprs
-                .iter()
-                .position(|(col, _)| col_idx.index + column_index_offset == col.index())
-        })
-        .collect()
-}
-
-/// Checks three conditions for pushing a projection down through a join:
-/// - Projection must narrow the join output schema.
-/// - Columns coming from left/right tables must be collected at the left/right
-///   sides of the output table.
-/// - Left or right table is not lost after the projection.
-fn join_allows_pushdown(
-    projection_as_columns: &[(Column, String)],
-    join_schema: &SchemaRef,
-    far_right_left_col_ind: i32,
-    far_left_right_col_ind: i32,
-) -> bool {
-    // Projection must narrow the join output:
-    projection_as_columns.len() < join_schema.fields().len()
-    // Are the columns from different tables mixed?
-    && (far_right_left_col_ind + 1 == far_left_right_col_ind)
-    // Left or right table is not lost after the projection.
-    && far_right_left_col_ind >= 0
-    && far_left_right_col_ind < projection_as_columns.len() as i32
-}
-
-/// If pushing down the projection over this join's children seems possible,
-/// this function constructs the new [`ProjectionExec`]s that will come on top
-/// of the original children of the join.
-fn new_join_children(
-    projection_as_columns: &[(Column, String)],
-    far_right_left_col_ind: i32,
-    far_left_right_col_ind: i32,
-    left_child: &Arc<dyn ExecutionPlan>,
-    right_child: &Arc<dyn ExecutionPlan>,
-) -> Result<(ProjectionExec, ProjectionExec)> {
-    let new_left = ProjectionExec::try_new(
-        projection_as_columns[0..=far_right_left_col_ind as _]
-            .iter()
-            .map(|(col, alias)| {
-                (
-                    Arc::new(Column::new(col.name(), col.index())) as _,
-                    alias.clone(),
-                )
-            })
-            .collect_vec(),
-        Arc::clone(left_child),
-    )?;
-    let left_size = left_child.schema().fields().len() as i32;
-    let new_right = ProjectionExec::try_new(
-        projection_as_columns[far_left_right_col_ind as _..]
-            .iter()
-            .map(|(col, alias)| {
-                (
-                    Arc::new(Column::new(
-                        col.name(),
-                        // Align projected expressions coming from the right
-                        // table with the new right child projection:
-                        (col.index() as i32 - left_size) as _,
-                    )) as _,
-                    alias.clone(),
-                )
-            })
-            .collect_vec(),
-        Arc::clone(right_child),
-    )?;
-
-    Ok((new_left, new_right))
 }
 
 #[cfg(test)]
@@ -999,6 +188,9 @@ mod tests {
     };
     use datafusion_physical_expr_common::sort_expr::{LexOrdering, LexRequirement};
     use datafusion_physical_optimizer::output_requirements::OutputRequirementExec;
+    use datafusion_physical_plan::coalesce_partitions::CoalescePartitionsExec;
+    use datafusion_physical_plan::filter::FilterExec;
+    use datafusion_physical_plan::joins::utils::{ColumnIndex, JoinFilter};
     use datafusion_physical_plan::streaming::StreamingTableExec;
     use datafusion_physical_plan::union::UnionExec;
 
@@ -1006,10 +198,12 @@ mod tests {
     use crate::datasource::listing::PartitionedFile;
     use crate::datasource::physical_plan::FileScanConfig;
     use crate::physical_plan::get_plan_string;
-    use crate::physical_plan::joins::StreamJoinPartitionMode;
+    use crate::physical_plan::joins::{
+        HashJoinExec, NestedLoopJoinExec, StreamJoinPartitionMode, SymmetricHashJoinExec,
+    };
 
-    use arrow_schema::{DataType, Field, Schema, SortOptions};
-    use datafusion_common::{JoinType, ScalarValue};
+    use arrow_schema::{DataType, Field, Schema, SchemaRef, SortOptions};
+    use datafusion_common::{JoinSide, JoinType, ScalarValue};
     use datafusion_execution::object_store::ObjectStoreUrl;
     use datafusion_execution::{SendableRecordBatchStream, TaskContext};
     use datafusion_expr::{
@@ -1021,6 +215,8 @@ mod tests {
     use datafusion_physical_expr::ScalarFunctionExpr;
     use datafusion_physical_plan::joins::PartitionMode;
     use datafusion_physical_plan::streaming::PartitionStream;
+
+    use itertools::Itertools;
 
     /// Mocked UDF
     #[derive(Debug)]
@@ -1324,61 +520,6 @@ mod tests {
                 .unwrap()
                 .eq(&expected_expr));
         }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_join_table_borders() -> Result<()> {
-        let projections = vec![
-            (Column::new("b", 1), "b".to_owned()),
-            (Column::new("c", 2), "c".to_owned()),
-            (Column::new("e", 4), "e".to_owned()),
-            (Column::new("d", 3), "d".to_owned()),
-            (Column::new("c", 2), "c".to_owned()),
-            (Column::new("f", 5), "f".to_owned()),
-            (Column::new("h", 7), "h".to_owned()),
-            (Column::new("g", 6), "g".to_owned()),
-        ];
-        let left_table_column_count = 5;
-        assert_eq!(
-            join_table_borders(left_table_column_count, &projections),
-            (4, 5)
-        );
-
-        let left_table_column_count = 8;
-        assert_eq!(
-            join_table_borders(left_table_column_count, &projections),
-            (7, 8)
-        );
-
-        let left_table_column_count = 1;
-        assert_eq!(
-            join_table_borders(left_table_column_count, &projections),
-            (-1, 0)
-        );
-
-        let projections = vec![
-            (Column::new("a", 0), "a".to_owned()),
-            (Column::new("b", 1), "b".to_owned()),
-            (Column::new("d", 3), "d".to_owned()),
-            (Column::new("g", 6), "g".to_owned()),
-            (Column::new("e", 4), "e".to_owned()),
-            (Column::new("f", 5), "f".to_owned()),
-            (Column::new("e", 4), "e".to_owned()),
-            (Column::new("h", 7), "h".to_owned()),
-        ];
-        let left_table_column_count = 5;
-        assert_eq!(
-            join_table_borders(left_table_column_count, &projections),
-            (2, 7)
-        );
-
-        let left_table_column_count = 7;
-        assert_eq!(
-            join_table_borders(left_table_column_count, &projections),
-            (6, 7)
-        );
 
         Ok(())
     }
@@ -2054,22 +1195,6 @@ mod tests {
             "    CsvExec: file_groups={1 group: [[x]]}, projection=[a, b, c, d, e], has_header=false"
             ];
         assert_eq!(get_plan_string(&after_optimize), expected);
-        Ok(())
-    }
-
-    #[test]
-    fn test_collect_column_indices() -> Result<()> {
-        let expr = Arc::new(BinaryExpr::new(
-            Arc::new(Column::new("b", 7)),
-            Operator::Minus,
-            Arc::new(BinaryExpr::new(
-                Arc::new(Literal::new(ScalarValue::Int32(Some(1)))),
-                Operator::Plus,
-                Arc::new(Column::new("a", 1)),
-            )),
-        ));
-        let column_indices = collect_column_indices(&[(expr, "b-(1+a)".to_string())]);
-        assert_eq!(column_indices, vec![1, 7]);
         Ok(())
     }
