+++ datafusion/physical-plan/src/joins/nested_loop_join.rs
@@ -37,6 +37,9 @@ use crate::joins::utils::{
     BuildProbeJoinMetrics, ColumnIndex, JoinFilter, OnceAsync, OnceFut,
 };
 use crate::metrics::{ExecutionPlanMetricsSet, MetricsSet};
+use crate::projection_utils::{
+    try_embed_projection, try_pushdown_through_join, EmbeddedProjection, JoinData,
+};
 use crate::{
     handle_state, DisplayAs, DisplayFormatType, Distribution, ExecutionPlan,
     ExecutionPlanProperties, PlanProperties, RecordBatchStream,
@@ -555,6 +558,44 @@ impl ExecutionPlan for NestedLoopJoinExec {
             &self.join_schema,
         )
     }
+
+    /// Tries to push `projection` down through `nested_loop_join`. If possible, performs the
+    /// pushdown and returns a new [`NestedLoopJoinExec`] as the top plan which has projections
+    /// as its children. Otherwise, returns `None`.
+    fn try_swapping_with_projection(
+        &self,
+        projection: &crate::projection::ProjectionExec,
+    ) -> Result<Option<Arc<dyn ExecutionPlan>>> {
+        // TODO: currently if there is projection in NestedLoopJoinExec, we can't push down projection to left or right input. Maybe we can pushdown the mixed projection later.
+        if self.contains_projection() {
+            return Ok(None);
+        }
+
+        if let Some(JoinData {
+            projected_left_child,
+            projected_right_child,
+            join_filter,
+            ..
+        }) = try_pushdown_through_join(
+            projection,
+            self.left(),
+            self.right(),
+            &[],
+            self.schema(),
+            self.filter(),
+        )? {
+            Ok(Some(Arc::new(NestedLoopJoinExec::try_new(
+                Arc::new(projected_left_child),
+                Arc::new(projected_right_child),
+                join_filter,
+                self.join_type(),
+                // Returned early if projection is not None
+                None,
+            )?)))
+        } else {
+            try_embed_projection(projection, self)
+        }
+    }
 }
 
 /// Asynchronously collect input into a single batch, and creates `JoinLeftData` from it
@@ -976,6 +1017,12 @@ impl<T: BatchTransformer + Unpin + Send> Stream for NestedLoopJoinStream<T> {
 impl<T: BatchTransformer + Unpin + Send> RecordBatchStream for NestedLoopJoinStream<T> {
     fn schema(&self) -> SchemaRef {
         Arc::clone(&self.schema)
+    }
+}
+
+impl EmbeddedProjection for NestedLoopJoinExec {
+    fn with_projection(&self, projection: Option<Vec<usize>>) -> Result<Self> {
+        self.with_projection(projection)
     }
 }
