+++ datafusion/physical-plan/src/joins/symmetric_hash_join.rs
@@ -47,6 +47,10 @@ use crate::joins::utils::{
     BatchTransformer, ColumnIndex, JoinFilter, JoinHashMapType, JoinOn, JoinOnRef,
     NoopBatchTransformer, StatefulStreamResult,
 };
+use crate::projection_utils::{
+    join_allows_pushdown, join_table_borders, new_join_children,
+    physical_to_column_exprs, update_join_filter, update_join_on,
+};
 use crate::{
     joins::StreamJoinPartitionMode,
     metrics::{ExecutionPlanMetricsSet, MetricsSet},
@@ -555,6 +559,81 @@ impl ExecutionPlan for SymmetricHashJoinExec {
                 batch_transformer: NoopBatchTransformer::new(),
             }))
         }
+    }
+
+    /// Tries to swap the projection with its input [`SymmetricHashJoinExec`]. If it can be done,
+    /// it returns the new swapped version having the [`SymmetricHashJoinExec`] as the top plan.
+    /// Otherwise, it returns None.
+    fn try_swapping_with_projection(
+        &self,
+        projection: &crate::projection::ProjectionExec,
+    ) -> Result<Option<Arc<dyn ExecutionPlan>>> {
+        // Convert projected PhysicalExpr's to columns. If not possible, we cannot proceed.
+        let Some(projection_as_columns) = physical_to_column_exprs(projection.expr())
+        else {
+            return Ok(None);
+        };
+
+        let (far_right_left_col_ind, far_left_right_col_ind) = join_table_borders(
+            self.left().schema().fields().len(),
+            &projection_as_columns,
+        );
+
+        if !join_allows_pushdown(
+            &projection_as_columns,
+            &self.schema(),
+            far_right_left_col_ind,
+            far_left_right_col_ind,
+        ) {
+            return Ok(None);
+        }
+
+        let Some(new_on) = update_join_on(
+            &projection_as_columns[0..=far_right_left_col_ind as _],
+            &projection_as_columns[far_left_right_col_ind as _..],
+            self.on(),
+            self.left().schema().fields().len(),
+        ) else {
+            return Ok(None);
+        };
+
+        let new_filter = if let Some(filter) = self.filter() {
+            match update_join_filter(
+                &projection_as_columns[0..=far_right_left_col_ind as _],
+                &projection_as_columns[far_left_right_col_ind as _..],
+                filter,
+                self.left().schema().fields().len(),
+            ) {
+                Some(updated_filter) => Some(updated_filter),
+                None => return Ok(None),
+            }
+        } else {
+            None
+        };
+
+        let (new_left, new_right) = new_join_children(
+            &projection_as_columns,
+            far_right_left_col_ind,
+            far_left_right_col_ind,
+            self.left(),
+            self.right(),
+        )?;
+
+        Ok(Some(Arc::new(SymmetricHashJoinExec::try_new(
+            Arc::new(new_left),
+            Arc::new(new_right),
+            new_on,
+            new_filter,
+            self.join_type(),
+            self.null_equals_null(),
+            self.right()
+                .output_ordering()
+                .map(|p| LexOrdering::new(p.to_vec())),
+            self.left()
+                .output_ordering()
+                .map(|p| LexOrdering::new(p.to_vec())),
+            self.partition_mode(),
+        )?)))
     }
 }
